New Mexfile PsychSound.mexmac -- OpenAL sound support =====================================================
This is the initial checkin of a reimplementation of
Psychtoolbox sound functions from scratch.

The new mexfile is called PsychSound.mexmac, its source
code is in Source/Common/PsychSound/

PsychSound is based on the free OpenAL audio library, the
sound counterpart to OpenGL. OpenAL is very powerful
and it is portable across Operating systems.

-> This should allow to easily port PsychSound to
   M$-Windoze and Linux when the time has come for
   a port.

-> Should also allow to take advantage of 3D sound support
   and other advanced features of OpenAL.

The current version doesn't use Apples OpenAL.framework
as this is not decent enough for our purpose. We use
a free LGPL'd version, contained in the ZIP file
OpenALSourceCode.zip which provides support for
audio capture.

PsychSound currently is in a very rough state. The basic
integration into PTB's Scriptingglue and such is there.

Support for audio capture via audio capture devices like
built-in microphone of iMacs or Line-In of desktop Macs,
or WebCam microphones is available and performs well.

This will allow for verbal response collection and
voice triggers. See the very unpolished/draft demos
SoundRecordingtest.m and SoundTrigger.m for basic usage.

Sound playback is not yet implemented and the sound capture
also needs significant refinement before this can go
into the Stable release.

-> Code is currently in successfull real world use by one
user (Sandhitsu Das from University of Pensylvania)...

Backup commit in preparation for future work, just to
get this into the official tree.

-Mario Kleiner