* Improved online registration:
Added some free software utility (GPL-License) "macid.exe" for
Microsoft Windows and modified the registration routine to use
it for MAC address query on Windows. Could not compile the MAC
code into Screen due to lack of needed system libraries. The
PsychContributed folder contains the utilities and their source
code + license info.

The registration routine now outputs its string in kind of
HTML tag style to simplify parsing and extension of the string
in the future. It also works on old Matlab-5 installs now and
is more robust.

* Screen.dll updated to lastest state of source code.

Guess we are nearly there for online registration.

-> Safe for beta.