PSYCHTOOLBOX 3.0.7 RELEASED
This is identical to the stable branch as of 6th October 2006.
See file ReleaseNotes.m in the Psychtoolbox folder for a short
summary of changes wrt. Psychtoolbox 1.0.6.

PTB is a pure backup version of stable, not seriously recommended
for anybody. Will be superseded in the next days by a massive update
to the 'stable' branch.

-Mario Kleiner