* Small fixes to libDC1394 video capture engine for small issues found   during testing on PowerPC. Plus some additional debug output in high   debug levels (>5).
  - Small bugfix in deferred start code (wait until start deadline).
  - Small bugfix to dropped frame accounting code.

* All modified MEX files for OS/X PowerPC rebuilt.

=> This one works high-perf and reliable on OS/X PowerPC.

Outstanding issues: Output of debug messages would crash, because
fprintf, printf statements inside libdc go to stdout/stderr, which
is a nono on Matlab. Need to add our own logging handlers and outstreams...

YUV->RGB conversion routine is very compute intense -- Too slow for
realtime 30 fps 640x480 color capture if conversion needs to be done
on a G5 1.6 Ghz. Don't know if this is because it is, or if the
conversion code is not optimized enough...