PsychDataPixx: Handle GPUs with more than 256 LUT slots correctly.
The 'CheckGPUSanity' routine which tries to correct problems with
non-identity GPU gamma table luts, display dithering etc. assumed
that the GPU always returns and requires a 256 rows LUT. This is
not true anymore for some high-end GPUs, e.g., the NVidia QuadroFX-3800
under Linux, which has a 2045 rows LUT.

This change makes sure the routine can cope with this now.