Screen/Linux: Try to do without direct linking against libraw1394.
Untested so far. We may not need to explicitely link against
libraw1394, but leave it to libdc1394 to define the implicit
dependency on it. This may help to avoid some of the problems
caused by bad software engineering at Mathworks side.

Build scripts updated to remove -lraw1394

Screen for 32 bit Matlab & Octave rebuilt, but not yet tested
with a Firewire camera.