Screen/GStreamer: Improve robustness of looped video playback.
A second method of rewinding the movie for looped playback
is now implemented. This one seems to be more robust against
slightly broken movie files, but seems to have a slightly
higher delay for rewinding on some odd movies.

The specific value of the non-zero Screen('PlayMovie', ..., loop, ...);
loop flag now decides on a case by case basis which method to use.

0 = Singleshot, 1 = New method, 2 = Old method, 3 = Both at the
same time for increased robustness or increased chaos.

-> Screen for 32-Bit Octave on Linux rebuilt and tested.
-> Now happily loops DualDiscs.mov, our worst movie.