PsychHID/Linux: Fix bug/regression in PsychHID('KbCheck', [], keylist)
A tiny mistake in the new KbCheck code for Linux causes KbCheck to totally
screw up its results iff and only iff both of the following is true:

a) The default system keyboard [] is queried.
b) A keylist of enabled keys is passed, e.g., as set via RestrictKeysForKbCheck().

This should fix it. Recompile and test pending.