Screen('MakeTexture'): Add new fast-path for planar texture creation.
If the optional 'specialFlags' flag is set to 4 and the GPU supports
FBO's and rectangle textures and shaders, and the input image is
sufficiently small (size * channels < max_texture_size of GPU), then
we try to store the texture data of multi-channel LA, RGB or RGBA
textures in a single, single channel Luminance texture.

Texel components are not stored pixel-interleaved as usual, but
in a planar fashion, the data blocks for R, G, B, A back-to-back
concatenated to each other. This way, texture creation can happen
without the need for copies at all (zero-copy) for uint8->uint8
textures, and without need for at least planar->interleaved
conversion for double input formats, avoiding the potential
cache trashing and memory controller bank-switching and other
nastyness. A GLSL shader needs to convert planar->interleaved
at 'DrawTexture' time to handle this format which is not supported
in hardware. Also, on use of such a texture as offscreen window or
in 'TransformTexture' or in normalization of its orientation, the
normalize routine must perform a interleaving blit to convert
the texture from planar to interleaved on the fly.

The format suffers certain restrictions when scrolling or
on borders. Therefore it is not enabled by default, but a
opt-in via specialFlags == 4.

Preliminary benchmark shows speedups of MakeTexture + DrawTexture
of anywhere between 2x (Luminance), 4x (RGBA), and 12x - 17x for
LA or RGB textures on the Mobility Radeon X1600. Not bad, at
least for this test case n=1.

-> MakeTextureTimingTest2 refined to test this new path.
-> Screen for OS/X Intel + Matlab rebuilt & tested.