PsychHID: Support for event buffers for KbQueues.
At least the core and OS/X implementation so far.

Keyboard queue event buffers buffer all key events of
a given keyboard queue. Events can be fetched later via
the new function PsychHID('KbQueueGetEvent'). Each event
contains timestamp, if it was a button press- or release event,
which raw keycode was affected, and the translated character
for that key. Translation of keys into characters is not yet
implemented, it returns -1 for any key.

Tested and shown to work on OSX, PsychHID for Matlab on
OSX Intel rebuilt.

Compile and test on Windows / Linux pending.
Demos, tests, key -> character translation pending.