Screen: Implement specialflag setting 2 for Quicktime movie playback.
The setting 2 is supposed to disable audio decoding and output
during movie playback. The brain-damaged Quicktime implementation
of Apple doesn't really allow us to disable decoding, voiding any
performance advantage we would get. Therefore we just mute the
audio volume on Quicktime for OS/X and Windows.

This is not as good as the GStreamer implementation which has
real advantages in performance and behaviour, but at least it
creates compatible behaviour at a superficial level.